requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlmodel>=0.0.8
sqlalchemy>=1.4
passlib[bcrypt]>=1.7
python-jose>=3.3.0
pytest>=7.0
httpx>=0.24
pytest-asyncio>=0.21.0
